VSCODE-GIT/requirements.txt
streamlit==1.40.2
google-generativeai==0.8.3
python-dotenv==1.0.0